hide history by default
