+++ app/helpers/follow_ups_helper.rb
@@ -65,4 +65,18 @@ module FollowUpsHelper
   def postal_invitation_disabled_for?(user, follow_up, department)
     !user.address? || user.archived_in?(department) || follow_up.rdv_pending? || follow_up.closed?
   end
+
+  def invitation_dates_by_format(invitations, invitation_formats)
+    invitation_dates_by_formats = invitation_formats.index_with { |_invitation_format| [] }
+    invitation_dates_by_formats.merge!(
+      invitations.group_by(&:format)
+                 .select { |format| invitation_formats.include?(format) }
+                 .transform_values { |invites| invites.map(&:created_at).sort.reverse }.to_h
+    )
+    invitation_dates_by_formats
+  end
+
+  def max_invitation_dates_for_a_format(invitation_dates_by_formats)
+    invitation_dates_by_formats.values.map(&:count).max
+  end
 end
