+++ app/helpers/users_helper.rb
@@ -11,16 +11,6 @@ module UsersHelper
     invitation_formats.include?(format)
   end
 
-  def invitation_dates_by_format(invitations, invitation_formats)
-    invitation_dates_by_formats = invitation_formats.index_with { |_invitation_format| [] }
-    invitation_dates_by_formats.merge!(
-      invitations.group_by(&:format)
-                 .delete_if { |format, _| !format.in?(invitation_formats) }
-                 .transform_values { |invites| invites.map(&:created_at).sort.reverse }.to_h
-    )
-    invitation_dates_by_formats
-  end
-
   def no_search_results?(users)
     users.empty? && params[:search_query].present?
   end
